renamed `_text_autocomplete` to `_text_addressit`
